Use float in Type::getNameOfType() instead of double

+++ compiler/codegen/Types.cpp
@@ -46,7 +46,7 @@ std::string Type::getNameOfType(TypeTag typeTag) {
     case TYPE_TAG_INT:
         return "int";
     case TYPE_TAG_FLOAT:
-        return "double";
+        return "float";
     case TYPE_TAG_STRING:
         return "string";
     case TYPE_TAG_BOOLEAN:
